Disable Provisioning when on Delphix
This is a simple patch to add a new provisioning class specific to
Delphix OS that disables the logic that would otherwise run. For now,
we don't intend to support the normal provisioning operations (adding
users, ssh keys, etc.), so we simply emit a log line stating the fact
that provisioning is disabled.