Disable Deprovisioning when on Delphix
This patch disables the default deprovisioning logic when running on a
Delphix OS system. It's likely that some but not all of the logic that
runs by default during deprovisoining will also need to be executed when
on a Delphix system, but for now, we opt to completely disable all of
the default actions. We can selectively enable the actions that we want
to run on Delphix in a later commit.