Add provisioning support and config file
This patch adds a new config file that is specific to when running on
Delphix OS. When the agent is installed on a Delphix system, this
Delphix specific configuration file will be copied and used to configure
the agent.

Additionally, the DelphixProvisionHandler class was added to support
provisioning when running on Delphix OS. While most of the normal
actions that occur during provisionging are disabled when on Delphix
(e.g. creating users, creating ssh keys, etc), configuring the system's
hostname is supported and should occur during this provisioning process.

And finally, some of the previous "logger.warn" messages have been
converted into "logger.error" messages. The idea being, any message that
is not expected to be emitted will use "logger.error", and any message
that is expected to be emitted will use "logger.warn".