AZ-231 Create post-push Jenkins job for WALinuxAgent repository
This adds a Jenkins job that is intended to be run "post push"; i.e. on
commits that have already been committed to the repository. This job
runs the "pre-push" job, which _should_ always pass, since this job
should have already been run on any code checked into the gate prior to
landing it. Additionally, if the "pre-push" job succeeds, the IPS
package will be updated for this repository using the new commit.