Add SMF manifest for new "waagent" service
This is a simple patch adding the manifest file defining a new "waagent"
SMF service. This will be automatically installed when installing the
agent on a Delphix system, but the service will be initially disabled.
Thus, for the service to actually do something useful, it will have to
be explicitly enabled by some other means.