Add commit hash information to package summary
This patch modifies the Jenkins job responsible for building the IPS
package, such that now the git commit hash information will be included
in the packages summary such that it can be easily retrievable using
"pkg info". As an example, the following package was generated using
this patch, and installed onto a Delphix system:

    $ pkg info walinuxagent
              Name: system/virtualization/walinuxagent
           Summary: Azure Linux Agent (g60e0360)
             State: Installed
         Publisher: delphix.com
           Version: 2.2.0.80
            Branch: 0.151014
    Packaging Date: December 22, 2016 03:14:09 AM
              Size: 1.12 MB
              FMRI: pkg://delphix.com/system/virtualization/walinuxagent@2.2.0.80-0.151014:20161222T031409Z

The part of the "Summary" enclosed in parenthesises encode the git
commit sha ("60e0360") of the git commit from which the package was
generated. This can help map a given package, to the exact version of
the code that was used to generate it.