Disable "set_admin_access_to_ip" on Delphix
The default behavior of the "set_admin_access_to_ip" function is to call
out to "iptables" to configure networking appropriately. Since Delphix
doesn't have "iptables", this function will report an error in the logs
when it's executed on a Delphix system.

Since it's uncertain if this function is required when the agent is
running on Delphix, it's simply been disabled. It can always be
re-enabled and implemented correctly for Delphix OS if it's determined
to be necessary.