Support SCSI timeout configuration on Delphix OS (fixed)
Previously, when comparing the existing timeout value to the desired
timeout value, this would always evaluate to "False" because the types
were different between the two variables; "current" is an integer, and
"timeout" is a string. Thus, we need to convert the two variables to the
same type, prior to attempting the comparison.