AZ-280 Investigate "-serialconsole" option
It turns out, the "-serialconsole" option is only available when using
the "waagent2.0" script. The "waagent" script doesn't support this
option, and this is the script that is used by the Delphix SMF service.

Rather than transition the Delphix SMF service to use the "waagent2.0"
script such that we can use this option, the provision code path of the
"waagent" script has been extended to support enablement of the serial
console.

To do this, a new "enable_serial_console" method was added to the
DefaultOSUtil class that is a no-op for all platforms except Delphix.
Additionally, this new method is then called from the provisioning code
path, such that the console will be enabled the first time the system
is provisioned.