Disable logging to "/dev/console" (again)
It turns out, that if we allow the agent to log it's messages to
"/dev/console", this pratically makes it impossible to use the console
interactively, which is useful when using our internal hyper-v systems.
Additionally, I'm not sure we actually want the console to be full of
messages from the Azure agent, even when running on the Azure cloud
platform where we don't have interactive console access.

Thus, I'm disabling the console logging functionality for now. I've left
the code in, with it commented out, so that it can serve as a reminder
for us to re-evaluate this decision later (e.g. if/when we decide to
upstream our changes).